+++ src/mulle-c/cmake/clobber/all/cmake/share/PostFiles.cmake
@@ -5,7 +5,9 @@ if( NOT __POST_FILES__CMAKE__)
       message( STATUS "# Include \"${CMAKE_CURRENT_LIST_FILE}\"" )
    endif()
 
-   include( AllLoad)
-   include( PostFilesCAux.cmake OPTIONAL)
+   include( UnwantedWarningsC)
+   include( CompilerDetectionC)
+
+   include( PostFilesCAux OPTIONAL)
 
 endif()