+++ src/mulle-c/cmake/clobber/all/cmake/share/PostLibrary.cmake
@@ -5,8 +5,8 @@ if( NOT __POST_LIBRARY__CMAKE__)
       message( STATUS "# Include \"${CMAKE_CURRENT_LIST_FILE}\"" )
    endif()
 
-   include( LinkManifest)
-   include( FinalOutput)
+   include( LinkManifestC)
+   include( FinalOutputC)
 
    include( PostLibraryCAux OPTIONAL)